requirements.txt
streamlit>=1.38.0
openai>=1.66.0
python-docx>=1.1.0
python-dotenv>=1.0.0
pandas>=1.4.0
requests>=2.31.0
aiohttp>=3.8.0
tqdm>=4.65.0
python-docx>=0.8.11